requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml -o requirements.txt
# Added manually: openai, requests for AI processing
anyio==4.6.2.post1
    # via openai
blinker==1.8.2
    # via flask
certifi==2024.8.30
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.0
    # via requests
distro==1.9.0
    # via openai
h11==0.14.0
    # via httpcore
httpcore==1.0.7
    # via httpx
httpx==0.27.2
    # via openai
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
openai==1.55.3
    # via rotary-phone-audio-guestbook (pyproject.toml)
pydantic==2.10.3
    # via openai
pydantic-core==2.27.1
    # via pydantic
requests==2.32.3
    # via rotary-phone-audio-guestbook (pyproject.toml)
sniffio==1.3.1
    # via
    #   anyio
    #   httpx
tqdm==4.67.1
    # via openai
typing-extensions==4.12.2
    # via
    #   openai
    #   pydantic
    #   pydantic-core
urllib3==2.2.3
    # via requests
cffi==1.17.1
    # via gevent
click==8.1.7
    # via flask
colorama==0.4.6
    # via click
colorzero==2.0
    # via gpiozero
flask==3.0.3
    # via rotary-phone-audio-guestbook (pyproject.toml)
gevent==24.11.1
    # via rotary-phone-audio-guestbook (pyproject.toml)
gpiozero==2.0.1
    # via rotary-phone-audio-guestbook (pyproject.toml)
greenlet==3.1.1
    # via gevent
gunicorn==23.0.0
    # via rotary-phone-audio-guestbook (pyproject.toml)
importlib-metadata==4.13.0
    # via
    #   flask
    #   gpiozero
importlib-resources==5.13.0
    # via gpiozero
itsdangerous==2.2.0
    # via flask
jinja2==3.1.4
    # via flask
markupsafe==2.1.5
    # via
    #   jinja2
    #   werkzeug
packaging==24.1
    # via gunicorn
psutil==7.0.0
    # via rotary-phone-audio-guestbook (pyproject.toml)
pycparser==2.22
    # via cffi
pyyaml==6.0.2
    # via rotary-phone-audio-guestbook (audioGuestBook.py)
rpi-gpio==0.7.1
    # via rotary-phone-audio-guestbook (audioGuestBook.py - Raspberry Pi GPIO)
adafruit-circuitpython-neopixel==6.3.11
    # via rotary-phone-audio-guestbook (audioGuestBook.py - LED strip control)
adafruit-blinka==8.47.0
    # via adafruit-circuitpython-neopixel (provides board module)
ruamel-yaml==0.18.6
    # via rotary-phone-audio-guestbook (pyproject.toml)
ruamel-yaml-clib==0.2.8
    # via ruamel-yaml
setuptools==74.1.2
    # via
    #   colorzero
    #   zope-event
    #   zope-interface
werkzeug==3.0.4
    # via flask
zipp==3.20.2
    # via
    #   importlib-metadata
    #   importlib-resources
zope-event==5.0
    # via gevent
zope-interface==7.2
    # via gevent
